UPD: merged 2 build scripts Bias calculation logic changes accoriding to todos Affects browsers/views/teplates
