+++ ligand/models.py
@@ -446,8 +446,8 @@ class BiasedExperimentAssay(models.Model):
     efficacy_sign = models.CharField(max_length=3, null=True)
     efficacy_unit = models.CharField(max_length=60, null=True)
     bias_reference = models.CharField(max_length=60, null=True)
-    bias_value = models.FloatField(max_length=60, null=True)
-    bias_value_initial = models.FloatField(max_length=60, null=True)
+    transduction_coef = models.FloatField(max_length=60, null=True)
+    relative_transduction_coef = models.FloatField(max_length=60, null=True)
 
     emax_ligand_reference = models.ForeignKey(Ligand, related_name = 'ExperimentAssay.bias_ligand_reference+',
                                         on_delete = models.CASCADE,
@@ -518,15 +518,12 @@ class AnalyzedAssay(models.Model):
     efficacy_measure_type = models.CharField(max_length=60, null=True)
     efficacy_sign = models.CharField(max_length=3, null=True)
     efficacy_unit = models.CharField(max_length=60, null=True)
-    potency = models.CharField(max_length=60, null=True)
-    t_coefficient = models.CharField(max_length=60, null=True)
-    t_value = models.CharField(max_length=60, null=True)
+    potency = models.FloatField(max_length=60, null=True)
+    transduction_coef = models.FloatField(max_length=60, null=True)
+    relative_transduction_coef = models.FloatField(max_length=60, null=True)
+    delta_relative_transduction_coef = models.FloatField(max_length=60, null=True)
     log_bias_factor = models.CharField(max_length=60, null=True)
-    log_bias_factor_a = models.CharField(max_length=60, null=True)
-    log_bias_factor_b = models.CharField(max_length=60, null=True)
-    log_bias_factor_c = models.CharField(max_length=60, null=True)
-    log_bias_factor_d = models.CharField(max_length=60, null=True)
-    log_bias_factor_e = models.CharField(max_length=60, null=True)
+    delta_emax_ec50 = models.FloatField(max_length=60, null=True)
     t_factor = models.CharField(max_length=60, null=True)
     assay_description = models.CharField(max_length=900, null=True)
     reference_ligand_id= models.CharField(max_length=900, null=True)
